Move confidence ellipsoid functions
